+++ packages/lib/src/tests.rs
@@ -37,7 +37,7 @@ macro_rules! create_tests {
 				.trim_end_matches(".luau")
 				.trim_end_matches(".lua")
 				.to_string();
-            let exit_code = lune.run(&script_name, &script).await?;
+            let exit_code = lune.try_run(&script_name, &script).await?;
             Ok(exit_code)
         }
     )* }